Reorganize some minor data types in package object
Also for easy importing of codecs
